背景追加

+++ database/migrations/2020_05_30_145916_create_parts_table.php
@@ -19,7 +19,7 @@ class CreatePartsTable extends Migration
             $table->integer('price'); //価格
             $table->integer('value'); //内容量
             $table->string('unit'); //単位
-            $table->float('bit'); //単価
+            $table->double('bit'); //単価
             $table->string('shop')->nullable()->default(null);
             $table->text('other')->nullable()->default(null);
             $table->integer('genru_id')->nullable()->default(null);
